+++ src/client/src/commons/menu/editor/actions/actionsInterfaces.ts
@@ -1,7 +1,7 @@
 export interface WorkgroupSelectProps {
   setState: React.Dispatch<React.SetStateAction<EditorSearchState>>;
   state: {
-    workgroup: Workgroup;
+    workgroup: number;
     enabledWorkgroups: Workgroup[];
   };
 }
@@ -15,12 +15,19 @@ export interface Workgroup {
 export interface ImportModalProps {
   setState: React.Dispatch<React.SetStateAction<EditorSearchState>>;
   state: {
+    selectedFile: Blob[] | null;
+  };
+}
+
+export interface ActionsModalProps {
+  setState: React.Dispatch<React.SetStateAction<EditorSearchState>>;
+  state: {
     creating: boolean;
     selectedLithologyFile: Blob[] | null;
     selectedBoreholeAttachments: Blob[] | null;
     selectedFile: Blob[] | null;
     upload: boolean;
-    workgroup: Workgroup;
+    workgroup: number;
     enabledWorkgroups: Workgroup[];
   };
   refresh: () => void;
